Extract message block helpers to standalone module so tests use real implementations
